Improved language check on vocab
